New two-step teacher sign up process & some upgrades
A new two-step teacher sign up process is introduced in this commit. From now on, a teacher can invite another teacher to join V4T using a simple form during their session in VSCode4Teaching. The app will return a link that can be shared with the new teacher, who will log in into the new web application and will be able to change the password as desired to start using the application.

- To introduce the new registration, a new frontend application based on Angular is created. This will allow, later on, to introduce all the features that would have to be implemented in a web application context.
- The Dockerfile is modified to facilitate the construction of the backend in Spring Boot, which introduces the Angular frontend as a previously built static resource.

In addition, Spring Boot has been upgraded to version 2.5.8 for security reasons (to fix vulnerabilities CVE-2021-44228, CVE-2021-45046 and CVE-2021-45105) and the H2 database dependency to version 2.1.210 (because of vulnerabilities CVE-2021-23463 and CVE-2022-23221).