New Dashboard preview mode
This commit introduces a new Dashboard preview mode specifically designed to allow opening the dashboard for any recorded exercise (and not only the currently downloaded exercise). In this mode, the ability to open exercise files and compare them with the original template is disabled (as these features require the exercise to be downloaded).