Exercise solution proposals (III) - New functionality for downloading solutions
This third commit introduces the ability to download solutions previously uploaded by teachers. To do so, enter:

- A new button is implemented in the Dashboard to allow teachers to publish the solutions they have added to their exercises.
- A new button is added in the bottom bar that allows students to download the solutions of exercises that have a solution and that is displayed only when it has been published by the teacher. When prespressed, it is downloaded to a new folder called solution in the main exercise directory.
- Teachers can view the solution they have attached to the exercises when downloading it, as from now on it is downloaded together with the template.

In addition, some bugs detected during the implementation of this functionality are fixed.

This commit is added to the previous and successive ones to complete the implemented functionality, being finalized for its incorporation in VSCode4Teaching v2.2.0.