Exercise solution proposals (IV) - Exercise editing lock after downloading solution
This fourth commit introduces the implementation of a new functionality to allow teachers to block or not the editing of exercises by students once the solution is downloaded.
To do so, it introduces:

- A new checkbox in the Dashboard that allows to enable or disable this feature to teachers. It can only be modified as long as the solution has not been made public.
- The fully implemented functionality of the button for downloading the solution by students is introduced in the exercises that allow it.

This commit is added to the previous and successive ones to complete the implemented functionality, being finalized for its incorporation in VSCode4Teaching v2.2.0.