Exercise solution proposals (V) - Comparison between student's exercise proposal and the teacher's solution
This commit introduces a new functionality that allows students, once they have downloaded the solution proposed by the teacher, to compare their proposal for solving the exercise with the solution predefined by the teacher.

- An algorithm for the comparison of two directories has been fully implemented. It is based on the generation of two trees representing both directories and their contents and on the combination of both trees with source memory.
- The user is shown by means of the native Visual Studio Code Quick Pick system the possibility to choose the file from which they want to visualize the differences.
- A new button is shown in the lower toolbar that allows to activate the functionality for displaying differences between directories. It is automatically displayed when downloading the solution.

The version of Visual Studio Code required to run the extension has been updated to 1.64.0 for the compatibility of the Quick Picks used.

This commit is added to the previous and successive ones to complete the implemented functionality, being finalized for its incorporation in VSCode4Teaching v2.2.0.