Exercise solution proposals (VI) - Improvement and introduction of new tests
This commit introduces new unit tests in both the server and client components, testing all new functionalities related to the introduction of teacher solution proposals and introducing, in addition, new tests of previously existing features. Some corrections to bugs detected during the test generation work are also introduced.

This commit is added to the previous ones to complete the implemented functionality, being finalized for its incorporation in VSCode4Teaching v2.2.0.