Bug fix in exercise addition
A bug has been detected in which the templates and solutions of the new exercises are not added correctly, since all the promises for the upload were introduced "all at once" as simultaneous requests to the server, which was saturated.

As a consequence, this is solved by sending the requests sequentially (one at a time).