Creation of a new Angular web application
A new updated Angular web application is introduced ready to serve as a scaffold on which new functionalities will be introduced in VSCode4Teaching.

Provisionally, the previous web application has been moved to the vscode4teaching-webapp-old directory. It will be removed when the migration process of the previously existing functionality is completed.