Introduces new dashboard screen
This commit introduces a new dashboard screen for users. It displays a list of courses and exercises available to the logged-in user.

In addition, it introduces a restructuring of the implemented services and new entities of the domain model, as well as the icons that will be used to indicate the status of the exercises and the existence or availability (if any) of their associated solution.