Add exercise synchronization for students - Step 1
This is the first of a set of commits aimed at implementing exercise synchronization functionality for students.

This functionality will have two implementations: in case of using a browser with the File System Access API, the algorithm will periodically (pseudo "real time") track file creations, modifications and deletions in the directory chosen by the user. Otherwise, the synchronization will be performed manually whenever the user wishes.

An ExerciseComponent component has been designed that will allow the synchronization to be executed in three steps (implemented as sub-components): the user starts by downloading his existing progress or the desired exercise template in ZIP format, unzips it and indicates in the browser where the resulting directory is and starts the routine as detailed above.

This commit introduces the first step of the above step enumeration. To do so, it introduces a graphical viewer with a progress bar that allows students to visualize the progress of the download and will display a window that allows them to choose the destination directory for the download.