Add exercise synchronization for students - Step 2
This is the second of a set of commits aimed at implementing exercise synchronization functionality for students.

The second step allows users to choose the directory in which the exercise content is included to start the synchronization.

Since there are two different implementations (depending on the availability of the File System Access API), a service has been implemented with the two implementations for reading directories, generating a common data structure that will later allow the content of the files to be read. It also includes an algorithm to determine the creations, modifications and deletions made between two of these structures.