Add exercise synchronization for students - Step 3
This is the third of a set of commits aimed at implementing exercise synchronization functionality for students.

The third step allows students, once they have chosen the local file system directory they wish to synchronize in step 2, to initiate automatic synchronization of the directory with the server (if the File System Access API is available), so that every 500 milliseconds the local directory will be scanned for new, deleted and modified files. Otherwise, the user must explicitly ask for the directory to be scanned (by manually selecting it), obtaining a list of new, edited and deleted files that is organized as a priority queue, assigning a priority for each synchronization job (one per file created, deleted or modified.)

Once the file list described above has been obtained, the creation, deletion and editing jobs are sent sequentially to the server using a synchronization algorithm that also allows the numeric progress to be reflected graphically so that the user can see on their own how it is evolving.

Previously, the application worked only with ZIP files. In order to work with the new individual file format, it has been necessary to adapt the server and introduce new endpoints that allow the exchange of individual files of each student's solution proposal.