Redesign of exercise synchronization for students
The user experience of the student exercise synchronization component has been completely redesigned.

From now on, students will select a local directory for each course in which they do exercises, and the application will manage and watch their contents.

Students can see on the screen all exercises according to their progress (not started, in progress and finished), and move between the different statuses. When an exercise is started or resumed, a directory is generated for each downloaded exercise with a predefined name and the previously stored progress or template is downloaded and unzipped inside. At this point, the changes in its files and subdirectories are automatically synchronized on the server.