Adds basic exercise progress tracking functionality for teachers
The Angular web application adds a new component hierarchy that allows teachers to access the details of their courses and view the exercises that comprise them. For each exercise, they can also view a screen with real-time synchronized student progress statistics and download the files of their proposals, as well as the original exercise template and its proposed solution (if any) on demand, incorporating it in the directory of their choice in their local file system.

To achieve this, the web application introduces some technical innovations, such as the refactoring of some components for the reuse of functionality or the generation of a Web Socket handler.

In addition, there have been introduced some other fixes:
- The way in which the paths of the modified files are saved in each proposed resolution has been modified, and it is not backwards-compatible.
- The implementation of the server to interpret encrypted authorization tokens has been completed.