Adds functionality to enable student self-enrollment in courses
Necessary UI elements are added to the interface to allow students to enroll in new courses from the sharing code provided by teachers.

In addition, the home page of the app is modified to show a customization in case students access a URL provided by a teacher to invite new people to a course.