Deletes extension and old web application (...)
The VSCode4Teaching web application already has the features of the extension, so this commit removes the entire implementation of the extension and the previous web application.
An update of the VSCode4Teaching documentation is pending, which will be done when the implementation of the new web application is finished.